+++ package.json
@@ -25,6 +25,7 @@
     "node": "10"
   },
   "dependencies": {
+    "prop-types": "15.7.2",
     "react": "^16.13.1",
     "react-dom": "16.13.1"
   }
