+++ src/components/app/app.jsx
@@ -1,0 +1,25 @@
+import React from "react";
+import PropTypes from "prop-types";
+import Main from "../main/main.jsx";
+
+
+const App = (props) => {
+  const {movieCard, films} = props;
+  return <div>
+    <Main
+      movieCard = {movieCard}
+      films = {films}
+    />
+  </div>;
+};
+
+App.propTypes = {
+  movieCard: PropTypes.shape({
+    name: PropTypes.string.isRequired,
+    genre: PropTypes.string.isRequired,
+    year: PropTypes.number.isRequired
+  }).isRequired,
+  films: PropTypes.array.isRequired,
+};
+
+export default App;
