+++ src/components/film-card/film-card.jsx
@@ -1,0 +1,23 @@
+import React from "react";
+import PropTypes from "prop-types";
+
+const FilmCard = (props) => {
+  const {films} = props;
+  return films.map((film, i) =>
+    <React.Fragment key = {film + i}>
+      <article className="small-movie-card catalog__movies-card">
+        <div className="small-movie-card__image">
+          <img src="img/fantastic-beasts-the-crimes-of-grindelwald.jpg" alt={film} width="280" height="175" />
+        </div>
+        <h3 className="small-movie-card__title">
+          <a className="small-movie-card__link" href="movie-page.html">{film}</a>
+        </h3>
+      </article>
+    </React.Fragment>);
+};
+
+FilmCard.propTypes = {
+  films: PropTypes.array.isRequired,
+};
+
+export default FilmCard;
